Added error checking

+++ src/consolecommands.h
@@ -35,6 +35,46 @@ void print_chance(string node) {
 			<< (g.get_node_infection_chance(node) * 100) << "%\n";
 }
 
+/** type checking */
+bool is_int(std::string str, int &val) {
+	try {
+		val = std::stoi(str); //don't call c_str()
+		return true;
+	} catch (std::exception const &e) {
+		return false;
+	}
+}
+
+bool is_double(std::string str, double &val) {
+	try {
+		val = std::stod(str); //don't call c_str()
+		return true;
+	} catch (std::exception const &e) {
+		return false;
+	}
+}
+
+bool is_float(std::string str, float &val) {
+	try {
+		val = std::stof(str); //don't call c_str()
+		return true;
+	} catch (std::exception const &e) {
+		return false;
+	}
+}
+
+bool is_valid_date(std::string &str) {
+	int n;
+	const char *c = str.c_str();
+	if (str.length() == 16)
+		return is_int(str.substr(0, 4), n) && (c[4] == '-') && (c[7] == '-')
+				&& (c[10] == '-') && (c[13] == ':')
+				&& is_int(str.substr(5, 2), n) && is_int(str.substr(8, 2), n)
+				&& is_int(str.substr(11, 2), n) && is_int(str.substr(14, 2), n);
+	else
+		return false;
+}
+
 string commands[] = { "help", "contact", "c", "node", "chance", "mark", "rule",
 		"save", "load", "printall", "delcontact", "d", "exit" };
 /** Description of commands.
@@ -90,15 +130,37 @@ void command_load(Bundle args) {
 }
 
 void command_rule(Bundle args) {
-	rule = rules[std::atoi(args[1].c_str())];
+	int n;
+	if (!is_int(args[1], n)) {
+		cout << "Invalid rule\n";
+		return;
+	}
+	if ((unsigned) n >= sizeof(rules) / sizeof(Rule)) {
+		cout << "Invalid index.\n";
+		return;
+	}
+	rule = rules[n];
 	// Before applying any rule, we must reset data first (infection 1 and 0)
 	g.apply_rule(rules[3]);
 	g.apply_rule(rule);
 }
 
 void command_mark(Bundle args) {
+	if (args.size != 3) {
+		cout << "Incorrect number of arguments. See help section.\n";
+		return;
+	}
+
+	int n;
+	if (!is_int(args[2], n)) {
+		cout << "Invalid value. 1 -> infected\t 0 -> clear\n";
+		return;
+	}
+
 	if (g.node_exists(args[1])) {
-		g.mark_infected(args[1], std::atoi(args[2].c_str()));
+		g.mark_infected(args[1], n);
+	} else {
+		cout << "Invalid node\n";
 	}
 	// Before applying any rule, we must reset data first (infection 1 and 0)
 	g.apply_rule(rules[3]);
@@ -107,27 +169,52 @@ void command_mark(Bundle args) {
 
 /** Shows or adds node to Graph */
 void command_chance(Bundle args) {
-	if (args.size == 2)
-		print_chance(args[1]);
+	if (args.size == 2) {
+		if (g.node_exists(args[1])) {
+			print_chance(args[1]);
+		} else {
+			cout << "Invalid node\n";
+		}
+	} else {
+		cout << "Incorrect number of arguments. See help section.\n";
+	}
+
 }
 
 /** Shows or adds node to Graph */
 void command_node(Bundle args) {
-	if (args.size == 3)
+	if (args.size == 3 || args.size == 2) {
+		int n = 0;
+		if (args.size == 3)
+			if (!is_int(args[2].c_str(), n)) {
+				cout
+						<< "Invalid value. Refer to manual for node types or leave empty\n";
+				return;
+			}
+
 		if (!g.node_exists(args[1])) {
-			g.add_node(Node::NodeFactory().set_id(args[1]).getNode());
+			g.add_node(
+					Node::NodeFactory().set_id(args[1]).set_type(n).getNode());
 		}
+	} else {
+		cout << "Incorrect number of arguments. See help section.\n";
+		return;
+	}
 }
 
 void command_delcontact(Bundle args) {
 	int argnum = args.size - 1;
 	if (argnum != 2) {
-		cout << "Incorrect number of arguments. Try again" << endl;
+		cout << "Incorrect number of arguments. See help section." << endl;
 		return;
 	}
 
 	if (argnum == 2) {
-		g.delete_contact(args[1], args[2]);
+		if (g.node_exists(args[1]) && g.node_exists(args[2])) {
+			g.delete_contact(args[1], args[2]);
+		} else {
+			cout << "Cannot delete contact with nonexistent node\n";
+		}
 	}
 
 }
@@ -143,14 +230,30 @@ void command_contact(Bundle args) {
 	if (argnum == 2) {
 		g.print_contact(args[1], args[2], cout);
 	} else {
+
+		float distance, exposure;
+
+		if (!is_float(args[3], distance)) {
+			cout << "Invalid distance\n";
+			return;
+		}
+		if (!is_float(args[4], exposure)) {
+			cout << "Invalid exposure\n";
+			return;
+		}
+		if (!is_valid_date(args[5])) {
+			cout << "Invalid date\n";
+			return;
+		}
+
 		if (!g.node_exists(args[1])) {
 			g.add_node(Node::NodeFactory().set_id(args[1]).getNode());
 		}
 		if (!g.node_exists(args[2])) {
 			g.add_node(Node::NodeFactory().set_id(args[2]).getNode());
 		}
-		g.add_contact(args[1], args[2], std::atof(args[3].c_str()),
-				std::atof(args[4].c_str()), args[5], rule);
+
+		g.add_contact(args[1], args[2], distance, exposure, args[5], rule);
 	}
 
 }
@@ -171,7 +274,8 @@ int findindex(string s) {
 }
 
 /** Array of function pointers for commands */
-void (*functions[])(Bundle args) = {command_help, command_contact, command_contact,
+void (*functions[])(
+		Bundle args) = {command_help, command_contact, command_contact,
 			command_node, command_chance, command_mark,
 			command_rule, command_save, command_load,
 			command_printall, command_delcontact, command_delcontact,
